requirements.txt
numpy==1.26.4
scipy==1.11.4
lightfm==1.17
pandas==2.2.2
psycopg2-binary==2.9.9
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
requests==2.32.3
streamlit==1.36.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
psycopg[binary]>=3.1.18
